+++ Algebra/PartialOrd.hs
@@ -18,11 +18,10 @@ module Algebra.PartialOrd (
     gfpFrom, unsafeGfpFrom
   ) where
 
-import qualified Data.Set as S
+import qualified Data.IntMap as IM
 import qualified Data.IntSet as IS
 import qualified Data.Map as M
-import qualified Data.IntMap as IM
-
+import qualified Data.Set as S
 
 -- | A partial ordering on sets
 -- (<http://en.wikipedia.org/wiki/Partially_ordered_set>) is a set equipped
@@ -102,10 +101,10 @@ instance PartialOrd IS.IntSet where
     leq = IS.isSubsetOf
 
 instance (Ord k, PartialOrd v) => PartialOrd (M.Map k v) where
-    m1 `leq` m2 = m1 `M.isSubmapOf` m2 && M.fold (\(x1, x2) b -> b && x1 `leq` x2) True (M.intersectionWith (,) m1 m2)
+    leq = M.isSubmapOfBy leq
 
 instance PartialOrd v => PartialOrd (IM.IntMap v) where
-    im1 `leq` im2 = im1 `IM.isSubmapOf` im2 && IM.fold (\(x1, x2) b -> b && x1 `leq` x2) True (IM.intersectionWith (,) im1 im2)
+    leq = IM.isSubmapOfBy leq
 
 instance (PartialOrd a, PartialOrd b) => PartialOrd (a, b) where
     -- NB: *not* a lexical ordering. This is because for some component partial orders, lexical